Laatste Styling PR dingetjes
Totale beschrijving staat bij de PR

+++ database/migrations/2013_03_29_105539_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
         {
     		$table->increments('id')->unique();
     		$table->string('name');
-    		$table->double('price');
+    		$table->decimal('price', 2);
     		$table->longText('description');
             $table->timestamps();
     	});
